+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed custom block explosion
+- Fixed custom max explosion radius
 
 
 ## [1.0.1] - 2025-2-12